Added top-level MPI logic to cortix_main

+++ cortix/src/cortix_main.py
@@ -2,12 +2,11 @@
 # -*- coding: utf-8 -*-
 # This file is part of the Cortix toolkit environment
 # https://cortix.org
-# Licensed under the University of Massachusetts Lowell LICENSE:
-# https://github.com/dpploy/cortix/blob/master/LICENSE.txt
 
 import os
 import shutil
 import logging
+from mpi4py import MPI
 from cortix.src.module import Module
 from cortix.src.utils.cortix_units import Units
 from cortix.src.utils.cortix_time import CortixTime
@@ -19,10 +18,14 @@ class Cortix:
 
     def __init__(self, work_dir="/tmp/"):
         self.work_dir = os.path.join(work_dir, 'cortix-wrk/')
+        self.comm = MPI.COMM_WORLD
+        self.rank = self.comm.Get_rank()
+        self.size = self.comm.size
 
-        # Create the work directory
-        shutil.rmtree(self.work_dir, ignore_errors=True)
-        os.mkdir(self.work_dir)
+        if self.rank == 0:
+            # Create the work directory
+            shutil.rmtree(self.work_dir, ignore_errors=True)
+            os.mkdir(self.work_dir)
 
         # Setup the global logger
         self.__create_logger()
@@ -32,7 +35,7 @@ class Cortix:
         self.evolve_time = CortixTime()
         self.time_step = CortixTime()
 
-        self.modules = list()
+        self.modules = []
 
         self.log.info('Created Cortix object')
 
@@ -40,12 +43,20 @@ class Cortix:
         assert isinstance(m, Module), "m must be a module"
         if m not in self.modules:
             self.modules.append(m)
+            m.rank = len(self.modules)
 
     def run(self):
         '''
-        Run the simulation
+        Run the Cortix simulation
         '''
-        pass
+        # Check for correct number of ranks
+        assert self.size == len(self.modules) + 1, "Incorrect nprocs"
+
+        # Run each module on its own rank 
+        for mod in self.modules:
+            if self.rank == mod.rank:
+                self.log.info("Launching Module {} on rank {}".format(mod, self.rank))
+                mod.run()
 
     def __create_logger(self):
         '''
@@ -61,7 +72,8 @@ class Cortix:
         console_handler.setLevel(logging.DEBUG)
 
         # Formatter added to handlers
-        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+        fs = '[{}] %(asctime)s - %(name)s - %(levelname)s - %(message)s'.format(self.rank)
+        formatter = logging.Formatter(fs)
         file_handler.setFormatter(formatter)
         console_handler.setFormatter(formatter)
 
@@ -70,4 +82,4 @@ class Cortix:
         self.log.addHandler(console_handler)
 
 if __name__ == "__main__":
-    c = Cortix("test")
+    c = Cortix()
